% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profile_details.R
\name{get_profile_metrics}
\alias{get_profile_metrics}
\title{Pull relevant metrics from a Duolingo user's profile}
\usage{
get_profile_metrics(username)
}
\arguments{
\item{username}{string, the username to search}
}
\value{
tibble, metrics
}
\description{
Pull relevant metrics from a Duolingo user's profile
}
